requirements.txt
aiogram>=3.3.0
aiohttp
qrcode
pillow
python-dotenv
